\name{blup}
\alias{blup}
\title{blup}
\usage{
data(blup)
}
\description{
Intercept estimations from a BLUP model of plant X fungal ectomycorrhizal interactions. The publication reporting these data in detail is forthcoming. The data come from the MycoDB project (Chaudhary et al., 2016; https://www.nature.com/articles/sdata201628).

}
\examples{
data(blup)

}
