\name{fungal_tree}
\alias{fungal_tree}
\title{fungal_tree}
\usage{
data(fungal_tree)
}
\description{
Example phylogenetic tree. The data come from the MycoDB project (Chaudhary et al., 2016; https://www.nature.com/articles/sdata201628). The tree is a result of phylogenetic reconstruction of all fungal species included in a meta-analysis that utilized a huge collection of controlled mycorrhizal interaction studies.

}
\examples{
data(fungal_tree)

}
