\name{fungal_tree_blup}
\alias{fungal_tree_blup}
\title{fungal_tree_blup}
\usage{
data(fungal_tree_blup)
}
\description{
Ectomycorrhizal fungi. The publication reporting these data in detail is forthcoming. The data come from the MycoDB project (Chaudhary et al., 2016; https://www.nature.com/articles/sdata201628).
}
\examples{
data(fungal_tree_blup)

}
