\name{mycor_ei}
\alias{mycor_ei}
\title{mycor_ei}
\usage{
data(mycor_ei)
}
\description{
Example dataset. The data come from the MycoDB project (Chaudhary et al., 2016; https://www.nature.com/articles/sdata201628). These ei values come from a phylogenetic meta-analysis of thousands of plants X mycorrhizal fungi combinations under controlled settings. The publication that results from this analysis is forthcoming.
}
\examples{
data(mycor_ei)

}
