\name{plant_tree}
\alias{plant_tree}
\title{plant_tree}
\usage{
data(plant_tree)
}
\description{
Example phylogenetic tree. The data come from the MycoDB project (Chaudhary et al., 2016; https://www.nature.com/articles/sdata201628). The tree is a result of phylogenetic reconstruction of all plant species included in a meta-analysis that utilized a huge collection of controlled mycorrhizal interaction studies.

}
\examples{
data(plant_tree)
}
