\name{plant_tree_blup}
\alias{plant_tree_blup}
\title{plant_tree_blup}
\usage{
data(plant_tree_blup)
}
\description{
Ectomycorrhizal plants. The publication reporting these data in detail is forthcoming. The data come from the MycoDB project (Chaudhary et al., 2016; https://www.nature.com/articles/sdata201628).

}
\examples{
data(plant_tree_blup)

}
