\name{rand_string}
\alias{rand_string}
\title{rand_string}
\usage{
rand_string(n = 1, len = 12)
}
\arguments{
\item{n}{number of strings to generate.}
\item{len}{ length of each string  }
}

\description{
Internal testing function for dualingTrees package. Not intended to be called directly.
}
\examples{
rand_string(10)
}
